RavenDB-22217 - Addressed PR Comments
