+++ src/Raven.Server/ServerWide/ClusterStateMachine.cs
@@ -4552,162 +4552,4 @@ namespace Raven.Server.ServerWide
             return false;
         }
     }
-
-    public class RachisLogIndexNotifications : AbstractRaftIndexNotifications
-    {
-        public Logger Log;
-
-        private readonly ConcurrentDictionary<long, TaskCompletionSource<object>> _tasksDictionary = 
-            new ConcurrentDictionary<long, TaskCompletionSource<object>>();
-
-        private readonly Queue<RecentLogIndexNotification> _recentNotifications = new Queue<RecentLogIndexNotification>();
-        
-        private SingleUseFlag _isDisposed = new SingleUseFlag();
-
-        public RachisLogIndexNotifications(CancellationToken token) : base(token)
-        {
-        }
-
-        public override void Dispose()
-        {
-            _isDisposed.Raise();
-            base.Dispose();
-            foreach (var task in _tasksDictionary.Values)
-            {
-                task.TrySetCanceled();
-            }
-        }
-
-        protected override string PrintLastNotifications()
-        {
-            var notifications = _recentNotifications.ToArray();
-            var builder = new StringBuilder(notifications.Length);
-            foreach (var notification in notifications)
-            {
-                builder
-                    .Append("Index: ")
-                    .Append(notification.Index)
-                    .Append(". Type: ")
-                    .Append(notification.Type)
-                    .Append(". ExecutionTime: ")
-                    .Append(notification.ExecutionTime)
-                    .Append(". Term: ")
-                    .Append(notification.Term)
-                    .Append(". LeaderErrorCount: ")
-                    .Append(notification.LeaderErrorCount)
-                    .Append(". LeaderShipDuration: ")
-                    .Append(notification.LeaderShipDuration)
-                    .Append(". Exception: ")
-                    .Append(notification.Exception)
-                    .AppendLine();
-            }
-            return builder.ToString();
-        }
-
-        public void RecordNotification(RecentLogIndexNotification notification)
-        {
-            _recentNotifications.Enqueue(notification);
-            while (_recentNotifications.Count > 25)
-                _recentNotifications.TryDequeue(out _);
-        }
-
-        public void SetTaskCompleted(long index, Exception e)
-        {
-            if (_tasksDictionary.TryGetValue(index, out var tcs))
-            {
-                // set the task as finished
-                if (e == null)
-                {
-                    if (tcs.TrySetResult(null) == false)
-                        LogFailureToSetTaskResult();
-                }
-                else
-                {
-                    if (tcs.TrySetException(e) == false)
-                        LogFailureToSetTaskResult();
-                }
-
-                void LogFailureToSetTaskResult()
-                {
-                    if (Log.IsInfoEnabled)
-                        Log.Info($"Failed to set result of task with index {index}");
-                }
-            }
-
-            _tasksDictionary.TryRemove(index, out _);
-        }
-
-        public void AddTask(long index)
-        {
-            Debug.Assert(_tasksDictionary.TryGetValue(index, out _) == false, $"{nameof(_tasksDictionary)} should not contain task with key {index}");
-            if (_isDisposed.IsRaised())
-                throw new ObjectDisposedException(nameof(RachisLogIndexNotifications));
-
-            _tasksDictionary.TryAdd(index, new TaskCompletionSource<object>(TaskCreationOptions.RunContinuationsAsynchronously));
-        }
-
-        public override async Task<bool> WaitForTaskCompletion(long index, Lazy<Task> waitingTask)
-        {
-            if (_tasksDictionary.TryGetValue(index, out var tcs) == false)
-            {
-                // the task has already completed
-                // let's check if we had errors in it
-                foreach (var error in _errors)
-                {
-                    if (error.Index == index)
-                        error.Exception.Throw(); // rethrow
-                }
-
-                return true;
-            }
-
-            var task = tcs.Task;
-
-            if (task.IsCompleted)
-            {
-                if (task.IsFaulted)
-                {
-                    try
-                    {
-                        await task; // will throw on error
-                    }
-                    catch (Exception e)
-                    {
-                        ThrowApplyException(index, e);
-                    }
-                }
-
-                if (task.IsCanceled)
-                    ThrowCanceledException(index, LastModifiedIndex);
-
-                return true;
-            }
-
-            var result = await Task.WhenAny(task, waitingTask.Value);
-
-            if (result.IsFaulted)
-                await result; // will throw
-
-            if (task.IsCanceled)
-                ThrowCanceledException(index, LastModifiedIndex);
-
-            if (result == task)
-                return true;
-
-            return false;
-        }
-
-    }
-
-    public class RecentLogIndexNotification
-    {
-        public string Type;
-        public TimeSpan ExecutionTime;
-        public long Index;
-        public int? LeaderErrorCount;
-        public long? Term;
-        public long? LeaderShipDuration;
-        public Exception Exception;
-    }
-
 }
